"Updated frontend components to support routes"
Several changes were made across all the frontend components to support routing with the 'react-router-dom' module. The 'DashboardCard' component was refactored to use props to pass necessary data, making it reusable and more flexible. Furthermore, routing was added to the sidebar and 'Dashboard' components to successfully link to different routes on click. 'Router.tsx' was also created to manage the incoming routes. These changes are significant as they enable user navigation on the application. Also, adjustments were made to 'public/index.html' to allow routing on the app. Lastly, package-lock.json was updated to include the needed dependencies for the changes.