Add state management to handle theme and active link setting
This commit introduces the useState Hook to handle theme and active link in the Sidebar and SidebarList components. Now localStorage is used to persist these settings between sessions. Also, the Link structure has changed to include dynamically set classNames and hrefs for better theme and active link management. A check around the initial theme setting has also been added to the Theme component, to default to 'dark' if none is present. In the Switch component, we've modified the label to be dynamic based on the current theme. Finally, some minor layout adjustments have been made in App and Sidebar components.