Add Leaflet package and remove @mui/material dependencies
Adjusted package dependencies to integrate the Leaflet package for map features and removed unnecessary @mui/material dependencies to optimize performance. Leaflet is a powerful open-source JavaScript library for mobile-friendly interactive maps. This change is necessary as part of adding geospatial features to our application. This will allow us to render visualizations on maps, which can help the user understand the data in a geographic context much better. Meanwhile, @mui/material dependencies have been removed as we don't use its components for the application.